Factor out AppPropsTestCase. Fix max BLOB test.
